# pokey: yet another performance enhancement