@Tafoid 2019-03-09
I have test cases where sound has regressed substantially:

centiped, missile (when shooting, the normal shot sound is no longer accurate)
indytemp (when using whip, it is barely audible and sounds different)
atetris, crystalc, foodf, jedi, mf_bdash, mhavoc, starwars and I'm sure many others  (music has lost mostly all pitch and is very low Hz.. sometimes it is merely static).