@schnitzeltony 2019-03-09
OK - I had tested some of those and did not hear any difference.

I need to check if it was caused by this patch because I sent other patches on pokey. Sorry for causing regression.