@schnitzeltony 2019-03-09
Just checked centipede: yes seems it was this patch. Will look into - if I don't find a fix soon I'll send a revert.