@Tafoid 2019-03-09
My testing is based on a binary pulled from f7ff691bd2
AFAIK this is the only pokey patch which was applied to master on March 8th.
My March 8th binary, based on 348a4ba53b , does not have the sound regressions.