@schnitzeltony 2019-03-09
Found it. It is the nasty rooky typo:

(m_AUDCTL & (CH1_HICLK || CH3_HICLK)) == 0
must be
(m_AUDCTL & (CH1_HICLK | CH3_HICLK)) == 0

Will do some tests:
* listen more carefully :)
* see if performance wins are still valid